Parts of code in instance.py is moved to instance_query_changes_worker.py and instance_query_worker.py. EimerDBInstance now inherits AbstractDbInstance. docstring should be removed from instance.py, but docstring-inheritance doesn't support Python 3.13 yet.
