+++ tests/test_instance_admin_user.py
@@ -84,7 +84,7 @@ class TestEimerDBInstanceAdminUser(TestEimerDBInstanceBase):
             "test_bucket/path/to/eimer/table1_changes/",
             format="parquet",
             partitioning="hive",
-            schema=self.instance._get_arrow_schema("table1", True),
+            schema=self.instance.get_arrow_schema("table1", True),
             filesystem=ANY,
         )
 
@@ -144,7 +144,7 @@ class TestEimerDBInstanceAdminUser(TestEimerDBInstanceBase):
             f"test_bucket/path/to/eimer/table1{suffix}/",
             format="parquet",
             partitioning="hive",
-            schema=self.instance._get_arrow_schema("table1", raw),
+            schema=self.instance.get_arrow_schema("table1", raw),
             filesystem=ANY,
         )
 
@@ -215,7 +215,7 @@ class TestEimerDBInstanceAdminUser(TestEimerDBInstanceBase):
             root_path="gs://test_bucket/path/to/eimer/table1_changes",
             partition_cols=None,
             basename_template="merged_commit_mocked_uuid_{i}.parquet",
-            schema=self.instance._get_arrow_schema("table1", True),
+            schema=self.instance.get_arrow_schema("table1", True),
             filesystem=ANY,
         )
         blob_1.delete.assert_called_once()
@@ -297,7 +297,7 @@ class TestEimerDBInstanceAdminUser(TestEimerDBInstanceBase):
             root_path=ANY,  # FIXME f"gs://test_bucket/path/to/eimer/table1{suffix}",
             partition_cols=None,
             basename_template="merged_commit_mocked_uuid_{i}.parquet",
-            schema=self.instance._get_arrow_schema("table1", raw),
+            schema=self.instance.get_arrow_schema("table1", raw),
             filesystem=ANY,
         )
